+++ src/sched/entry/coll/coll_entry_param.hpp
@@ -28,7 +28,7 @@ struct ccl_coll_entry_param {
     const size_t* recv_counts;
     ccl_datatype dtype;
     ccl::reduction reduction;
-    size_t root;
+    int root;
     ccl_comm* comm;
 
     ccl_coll_param to_coll_param() const {
